webpage: drop the stats_extras and webpage_extras tag libraries
stats_extras and webpage_extras were tag libraries shipped with the
infos module of apis-core. This module is now a separate module. The
webpage templates don't use the stats_extras functionality and the
webpage_extras functionality was simply to list projects and some
static content. Neither was not used, so we can simply drop the
templatetags.