feat: cleanup base template
The webpage base template copied major parts of the APIS `base.html`
template. To make things easier, this commit cleans up the template so
that we don't have to implement changes in two places.

This needs at least apis-core-rdf version 0.3.0, because it relies on
the `userlogin-menu` block which was implemented in 7bc4a29